try to improve folly/io/async/test:async_test - AsyncSocketTest.Write
Summary:
the test fails a lot as seen on the task.
To kill the connection in said test, we depeneded
on destructor, however this uses a delayed destruction
patter so I guess we did not destroy on time and
we were failing the tests.
I close socket explicitly. No idea if this misses
the point of whole test.

Reviewed By: yfeldblum

Differential Revision: D4171111

fbshipit-source-id: 01b8c97a04a2bcdfaeeea522ca8bcae343d6bb5d