correct the description about the thread to run
Summary:
exchange A and B in this sentence :
"Or, maybe x will execute in Thread B, but y and/or z will execute in Thread A".

I think x in B and y,z in A is impossible. but x in A and y,z in B is possible, in that example.
Closes https://github.com/facebook/folly/pull/514

Reviewed By: yfeldblum

Differential Revision: D4176766

Pulled By: Orvid

fbshipit-source-id: cf3a47a30d7e43e5291d6b6401198025beba33f8