remove override from adjacent_tokens_only
Summary:
Our internal version of boost contains a patch that reverts the following commit to boost filesystem module:
https://github.com/boostorg/program_options/commit/88dea3c6fdea8c9ea894911897b1770599c383e4

It was done due to a breaking change that disallows the use of "--key value" for implicit options, allowing only "--key=value" style syntax. Since there are too many tests to fix, and it's probably going to be reverted in the upstream as well (see below), removing the `override` keyword here to fix folly build.

https://github.com/boostorg/program_options/issues/25

Reviewed By: meyering

Differential Revision: D4415347

fbshipit-source-id: c215331338b3aa218d26890a83aa0b363514fe79