portability: handle libstdc++ redeclarations
The interfaces here are decorated with `__attribute__((__noreturn__))`
rather than the C++11 attribute `[[noreturn]]`.  The two are subtly
different in behaviour and thus do not compare equal.  The redeclaration
therefore is treated as changing the attributes (rather than just the
attribute spelling).  This will fail to compile as a result.